# Use Map and WeakMap to remove leaks and enhance perf