--- a/.travis.yml
+++ b/.travis.yml
@@ -1,8 +1,4 @@
 language: node_js
-dist: trusty
-before_script:
-  - npm install
-node_js: 6
+node_js: 8
 script:
-  - npm test
-  - npm run bench
+  - npm test && npm run bench