--- a/package.json
+++ b/package.json
@@ -1,6 +1,6 @@
 {
   "name": "fast-json-patch",
-  "version": "2.0.6",
+  "version": "2.0.7",
   "description": "Fast implementation of JSON-Patch (RFC-6902) with duplex (observe changes) capabilities",
   "homepage": "https://github.com/Starcounter-Jack/JSON-Patch",
   "keywords": [