Performance improvements: 
```shell
add op: 117% faster
remove op: 109% faster
replace: 112% faster
test: 131%
geneate: 110% faster
```

### Raw numbers
#### Before: 
```shell
add operation
 Ops/sec: 3347825 ±4.97% Ran 212855 times in 0.064 seconds.
remove operation
 Ops/sec: 2846822 ±2.96% Ran 165023 times in 0.058 seconds.
replace operation
 Ops/sec: 3490267 ±4.53% Ran 210019 times in 0.060 seconds.
move operation
 Ops/sec: 248869 ±3.06% Ran 14581 times in 0.059 seconds.
copy operation
 Ops/sec: 165951 ±4.60% Ran 18362 times in 0.111 seconds.
test operation
 Ops/sec: 556021 ±9.20% Ran 37831 times in 0.068 seconds.
```
#### After:
```shell
add operation
 Ops/sec: 3937701 ±1.55% Ran 220030 times in 0.056 seconds.
remove operation
 Ops/sec: 3124531 ±1.10% Ran 169656 times in 0.054 seconds.
replace operation
 Ops/sec: 3924373 ±2.95% Ran 218723 times in 0.056 seconds.
move operation
 Ops/sec: 273714 ±1.22% Ran 15114 times in 0.055 seconds.
copy operation
 Ops/sec: 202550 ±2.24% Ran 18524 times in 0.091 seconds.
test operation
 Ops/sec: 733608 ±0.97% Ran 39733 times in 0.054 seconds.
```

✔️ All Tests are passing in IE

Fixes https://github.com/Starcounter-Jack/JSON-Patch/issues/208