@warpech 2019-06-28
> ✔️ All Tests are passing in IE

This PR broke compatibility with IE versions lower than 11. I am not saying this is a disaster, but it was worth noting, especially since another part in code still claims to support IE8