@alshakero 2019-07-01
See https://github.com/Starcounter-Jack/JSON-Patch/issues/223